* commit 'cde526bdc1a543e9601257d149e99d036b78044f':
  fix automatic code generation from schema
  makefile update: mv instead cp
  was obsolete, redesigned to target the makefile in schema folder
  define schema variable as optional (default to 3.1)
  fix incorrect dependency
  docker update
  gitignore update
  add docker container
  add -o option
  fix warnings
  build system update: output location is set by cmake
  update to version 3.1.6
  version number changed to 3.1.5 (next release)
  npm packaging and publishing
  ostream operator << with indenter exported (windows)
  travis change: compile also lilypond branch

+++ src/guido/xml2guidovisitor.cpp
@@ -247,7 +247,7 @@ namespace MusicXML2
                 if (previousStaffHasLyrics)
                 {
                     tag2->add (guidoparam::create("dy=-5", false));
-                }else if (ps.fStaffDistances.size()> (targetStaff-1)) {
+                }else if (ps.fStaffDistances.size()> size_t(targetStaff-1)) {
                     
                     if (ps.fStaffDistances[targetStaff-1] > 0) {
                         float xmlDistance = ps.fStaffDistances[targetStaff-1] - 50.0;
